backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
PyJWT==2.1.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
mysql-connector-python==8.0.26
python-dotenv==0.19.0
stripe==2.60.0
Pillow==8.3.1
bcrypt==3.2.0
requests==2.26.0
pytest==6.2.5 